Implement first handshake
The first part of the client/server handshake is a periodic GET from the
client to a fictitious /style.css on the server. If the server is
running, it'll respond with a 200 OK and a random bs css file for the
appearance of legitimacy. The server response is loaded from file, so
the server script isn't standalone (that may change).