Improve dlexec to use tempfile
Use tempfile module to save executable. It creates a /much/ better
random location for the file. We need to create the NamedTemporaryFile
with a delete=False parameter because otherwise there seems to be this
weird race condition that occurs where the file doesn't actually get
executed because the file gets automatically deleted before it's
executed? Seems pretty odd to think that it would get deleted even
before the exec syscall would load it into memory, but I guess if the
parent doesn't have much to do and the child has a lot to do in between
the fork and exec, it could happen.