Move common socket code into poetsocket.py
Move PoetSocket class which was identical in both client and server into
separate poetsocket.py file which can be imported by the server.
We don't want to have multiple files to drop on the client though, so we
get around this by packaging poetsocket.py with __main__.py (ln to
client.py) into a zip file which can fed into the python interpreter and
executed as normal.