Add modindex.txt to package
This is so the client/server packages can know the names of the modules
to load, because otherwise they don't have any way to know. They aren't
able to use the filesystem because in production, they're executing in a
zip file and can't easily list the contents of the modules therein.