Move module decorator and client dict into module.py
Modules now `import module` and then use the @module.client_handler
decorator to define client handler function. This corresponds with a
change in client so that the client no longer iterates through the
self.mods, it just iterates through the module.client_commands which
collects all of the registered commands and handlers.